refactor: split tests to optimize workers
