+++ .github/workflows/test-frontend.yaml
@@ -80,7 +80,7 @@ jobs:
             soloRequired: true
             backendRequired: true
             sharedEnvironment: true
-            workers: 3
+            workers: 4
     steps:
       - name: Harden Runner
         uses: step-security/harden-runner@fe104658747b27e96e4f7e80cd0a94068e53901d # v2.16.1
