+++ automation/tests/organization-advanced/organizationTransactionExecutionTests.test.ts
@@ -1,0 +1,227 @@
+import { expect, Page, test } from '@playwright/test';
+import { OrganizationPage, UserDetails } from '../../pages/OrganizationPage.js';
+import { LoginPage } from '../../pages/LoginPage.js';
+import { TransactionPage } from '../../pages/TransactionPage.js';
+import { flushRateLimiter } from '../../utils/db/databaseUtil.js';
+import { setDialogMockState } from '../../utils/runtime/dialogMocks.js';
+import type { TransactionToolApp } from '../../utils/runtime/appSession.js';
+import { waitForValidStart } from '../../utils/runtime/timing.js';
+import { setupOrganizationAdvancedFixture } from '../helpers/fixtures/organizationAdvancedFixture.js';
+import {
+  setupOrganizationSuiteApp,
+  teardownOrganizationSuiteApp,
+} from '../helpers/bootstrap/organizationSuiteBootstrap.js';
+import type { ActivatedTestIsolationContext } from '../../utils/setup/sharedTestEnvironment.js';
+import { createSequentialOrganizationNicknameResolver } from '../helpers/support/organizationNamingSupport.js';
+
+let app: TransactionToolApp;
+let window: Page;
+let globalCredentials = { email: '', password: '' };
+
+let transactionPage: TransactionPage;
+let organizationPage: OrganizationPage;
+let loginPage: LoginPage;
+let isolationContext: ActivatedTestIsolationContext | null = null;
+let organizationNickname = 'Test Organization';
+
+let firstUser: UserDetails;
+let complexKeyAccountId: string;
+const resolveOrganizationNickname = createSequentialOrganizationNicknameResolver();
+
+test.describe('Organization Transaction execution-type tests @organization-advanced', () => {
+  test.slow();
+
+  test.beforeAll(async () => {
+    ({
+      app,
+      window,
+      loginPage,
+      transactionPage,
+      organizationPage,
+      isolationContext,
+    } = await setupOrganizationSuiteApp(test.info()));
+    window.on('console', msg => {
+      const text = msg.text();
+      if (
+        text.includes('[TXD-DBG]') ||
+        text.includes('[SIG-AUDIT-DBG]') ||
+        text.includes('[ORG-USER-DBG]')
+      ) {
+        console.log('[BROWSER]', text);
+      }
+    });
+  });
+
+  test.beforeEach(async ({}, testInfo) => {
+    await flushRateLimiter();
+    await setDialogMockState(window, { savePath: null, openPaths: [] });
+
+    organizationNickname = resolveOrganizationNickname(testInfo.title);
+    const fixture = await setupOrganizationAdvancedFixture(
+      window,
+      loginPage,
+      organizationPage,
+      organizationNickname,
+    );
+    globalCredentials.email = fixture.localCredentials.email;
+    globalCredentials.password = fixture.localCredentials.password;
+    firstUser = fixture.firstUser;
+    complexKeyAccountId = fixture.complexKeyAccountId;
+    await transactionPage.clickOnTransactionsMenuButton();
+
+    await organizationPage.waitForElementToDisappear('.v-toast__text');
+    await organizationPage.closeDraftModal();
+
+    if (process.env.CI) {
+      await new Promise(resolve => setTimeout(resolve, 1000));
+    }
+  });
+
+  test.afterEach(async () => {
+    try {
+      await organizationPage.logoutFromOrganization();
+    } catch {
+      // Several tests delete or fully consume the current org session.
+      // The next beforeEach recreates the fixture from scratch.
+    }
+  });
+
+  test.afterAll(async () => {
+    await teardownOrganizationSuiteApp(app, isolationContext);
+  });
+
+  test('Verify user can execute transfer transaction with complex account', async () => {
+    const { txId, validStart } = await organizationPage.transferAmountBetweenAccounts(
+      complexKeyAccountId,
+      '15',
+    );
+    await organizationPage.closeDraftModal();
+    await transactionPage.clickOnTransactionsMenuButton();
+    await organizationPage.logoutFromOrganization();
+
+    await organizationPage.logInAndSignTransactionByAllUsers(
+      globalCredentials.password,
+      txId ?? '',
+    );
+    await organizationPage.signInOrganization(
+      firstUser.email,
+      firstUser.password,
+      globalCredentials.password,
+    );
+    await waitForValidStart(validStart ?? '');
+
+    await organizationPage.clickOnHistoryTab();
+    const transactionDetails = await organizationPage.getHistoryTransactionDetails(txId ?? '');
+    expect(transactionDetails?.transactionId).toBe(txId);
+    expect(transactionDetails?.transactionType).toBe('Transfer');
+    expect(transactionDetails?.validStart).toBeTruthy();
+    expect(transactionDetails?.detailsButton).toBe(true);
+    expect(transactionDetails?.status).toBe('SUCCESS');
+  });
+
+  test('Verify user can execute approve allowance with complex account', async () => {
+    const { txId, validStart } = await organizationPage.approveAllowance(
+      complexKeyAccountId,
+      '10',
+    );
+    await organizationPage.closeDraftModal();
+    await transactionPage.clickOnTransactionsMenuButton();
+    await organizationPage.logoutFromOrganization();
+
+    await organizationPage.logInAndSignTransactionByAllUsers(
+      globalCredentials.password,
+      txId ?? '',
+    );
+    await organizationPage.signInOrganization(
+      firstUser.email,
+      firstUser.password,
+      globalCredentials.password,
+    );
+    await waitForValidStart(validStart ?? '');
+
+    await organizationPage.clickOnHistoryTab();
+    const transactionDetails = await organizationPage.getHistoryTransactionDetails(txId ?? '');
+    expect(transactionDetails?.transactionId).toBe(txId);
+    expect(transactionDetails?.transactionType).toBe('Account Allowance Approve');
+    expect(transactionDetails?.validStart).toBeTruthy();
+    expect(transactionDetails?.detailsButton).toBe(true);
+    expect(transactionDetails?.status).toBe('SUCCESS');
+  });
+
+  test.skip('Verify user can execute file create with complex account', async () => {
+    const { txId } = await organizationPage.ensureComplexFileExists(
+      complexKeyAccountId,
+      globalCredentials,
+      firstUser,
+      600,
+    );
+    const transactionDetails = await organizationPage.getHistoryTransactionDetails(txId ?? '');
+    expect(transactionDetails?.transactionId).toBe(txId);
+    expect(transactionDetails?.transactionType).toBe('File Create');
+    expect(transactionDetails?.validStart).toBeTruthy();
+    expect(transactionDetails?.detailsButton).toBe(true);
+    expect(transactionDetails?.status).toBe('SUCCESS');
+  });
+
+  test.skip('Verify user can execute file update with complex account', async () => {
+    const { fileId } = await organizationPage.ensureComplexFileExists(
+      complexKeyAccountId,
+      globalCredentials,
+      firstUser,
+      600,
+    );
+    const { txId, validStart } = await organizationPage.fileUpdate(
+      fileId ?? '',
+      complexKeyAccountId,
+      'newContent',
+    );
+    await organizationPage.closeDraftModal();
+    await organizationPage.signTxByAllUsersAndRefresh(globalCredentials, firstUser, txId ?? '');
+    await waitForValidStart(validStart ?? '');
+
+    const transactionDetails = await organizationPage.getHistoryTransactionDetails(txId ?? '');
+    expect(transactionDetails?.transactionId).toBe(txId);
+    expect(transactionDetails?.transactionType).toBe('File Update');
+    expect(transactionDetails?.validStart).toBeTruthy();
+    expect(transactionDetails?.detailsButton).toBe(true);
+    expect(transactionDetails?.status).toBe('SUCCESS');
+  });
+
+  test.skip('Verify user can execute file append with complex account', async () => {
+    const { fileId } = await organizationPage.ensureComplexFileExists(
+      complexKeyAccountId,
+      globalCredentials,
+      firstUser,
+      600,
+    );
+    const { txId, validStart } = await organizationPage.fileAppend(
+      fileId ?? '',
+      complexKeyAccountId,
+      'appendContent',
+    );
+    await organizationPage.closeDraftModal();
+    await organizationPage.signTxByAllUsersAndRefresh(globalCredentials, firstUser, txId ?? '');
+    await waitForValidStart(validStart ?? '');
+
+    const transactionDetails = await organizationPage.getHistoryTransactionDetails(txId ?? '');
+    expect(transactionDetails?.transactionId).toBe(txId);
+    expect(transactionDetails?.transactionType).toBe('File Append');
+    expect(transactionDetails?.validStart).toBeTruthy();
+    expect(transactionDetails?.detailsButton).toBe(true);
+    expect(transactionDetails?.status).toBe('SUCCESS');
+  });
+
+  test('Verify user can execute account delete with complex account', async () => {
+    const { txId, validStart } = await organizationPage.deleteAccount(complexKeyAccountId);
+    await organizationPage.closeDraftModal();
+    await organizationPage.signTxByAllUsersAndRefresh(globalCredentials, firstUser, txId ?? '');
+    await waitForValidStart(validStart ?? '');
+
+    const transactionDetails = await organizationPage.getHistoryTransactionDetails(txId ?? '');
+    expect(transactionDetails?.transactionId).toBe(txId);
+    expect(transactionDetails?.transactionType).toBe('Account Delete');
+    expect(transactionDetails?.validStart).toBeTruthy();
+    expect(transactionDetails?.detailsButton).toBe(true);
+    expect(transactionDetails?.status).toBe('SUCCESS');
+  });
+});
